Added verify_instances to utils.py to be called by both logs.py and status.py. Also moved the tests to test_utils.py
